fuse: parse socket control message using the standard library
make mount procedure cleaner by using standard library
to parse socker control message
